+++ src/lib_utils/cli_args.rs
@@ -73,3 +73,5 @@ impl CliArgs {
         Ok(())
     }
 }
+
+//TODO inroduce testing here